[0.3] fix: compiler warnings
git-svn-id: https://kdlibcpp.svn.codeplex.com/svn@87624 9b283d60-5439-405e-af05-b73fd8c4d996

+++ kdlib/source/stack.cpp
@@ -86,7 +86,7 @@ unsigned long StackFrame::getTypedParamCount()
 
     TypedVarPtr func = mod->getFunctionByAddr(m_ip);
 
-    return func->getElementCount();
+    return static_cast<unsigned long>(func->getElementCount());
 }
 
 ///////////////////////////////////////////////////////////////////////////////
